Move Base Builders to own namespace
