+++ DSharpPlus/Entities/Builders/Abstractions/MessageBuilder.cs
@@ -1,0 +1,116 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace DSharpPlus.Entities.Abstractions
+{
+    /// <summary>
+    /// Represents the message that will be Created or Modified.
+    /// </summary>
+    public abstract class MessageBuilder<T>
+    {
+        /// <summary>
+        /// Gets or Sets the Message to be sent.
+        /// </summary>
+        public string Content
+        {
+            get => this._content;
+            set
+            {
+                if (value != null && value.Length > 2000)
+                    throw new ArgumentException("Content cannot exceed 2000 characters.", nameof(value));
+                this._content = value;
+            }
+        }
+        private string _content;
+
+        /// <summary>
+        /// Gets or Sets the Embed to be sent.
+        /// </summary>
+        public DiscordEmbed Embed { get; set; }
+
+        /// <summary>
+        /// Gets or Sets if the message should be TTS.
+        /// </summary>
+        public bool IsTTS { get; set; } = false;
+
+        /// <summary>
+        /// Gets the Allowed Mentions for the message to be sent. 
+        /// </summary>
+        public List<IMention> Mentions { get; internal set; } = null;
+
+        /// <summary>
+        /// Sets the Content of the Message.
+        /// </summary>
+        /// <param name="content">The content to be set.</param>
+        /// <returns></returns>
+        public T WithContent(string content)
+        {
+            this.Content = content;
+            return (T)Convert.ChangeType(this, typeof(T));
+        }
+
+        /// <summary>
+        /// Sets if the message should be TTS.
+        /// </summary>
+        /// <param name="isTTS">If TTS should be set.</param>
+        /// <returns></returns>
+        public T HasTTS(bool isTTS)
+        {
+            this.IsTTS = isTTS;
+            return (T)Convert.ChangeType(this, typeof(T));
+        }
+
+        /// <summary>
+        /// Sets if the message will have an Embed.
+        /// </summary>
+        /// <param name="embed">The embed that should be sent.</param>
+        /// <returns></returns>
+        public T WithEmbed(DiscordEmbed embed)
+        {
+            this.Embed = embed;
+            return (T)Convert.ChangeType(this, typeof(T));
+        }
+
+        /// <summary>
+        /// Sets if the message has allowed mentions.
+        /// </summary>
+        /// <param name="allowedMention">The allowed Mention that should be sent.</param>
+        /// <returns></returns>
+        public T WithAllowedMention(IMention allowedMention)
+        {
+            if (this.Mentions != null)
+                this.Mentions.Add(allowedMention);
+            else
+                this.Mentions = new List<IMention> { allowedMention };
+
+            return (T)Convert.ChangeType(this, typeof(T));
+        }
+
+        /// <summary>
+        /// Sets if the message has allowed mentions.
+        /// </summary>
+        /// <param name="allowedMentions">The allowed Mentions that should be sent.</param>
+        /// <returns></returns>
+        public T WithAllowedMentions(IEnumerable<IMention> allowedMentions)
+        {
+            if (this.Mentions != null)
+                this.Mentions.AddRange(allowedMentions);
+            else
+                this.Mentions = allowedMentions.ToList();
+
+            return (T)Convert.ChangeType(this, typeof(T));
+        }
+
+        /// <summary>
+        /// Allows for clearing the Message Builder so that it can be used again to send a new message.
+        /// </summary>
+        public abstract void Clear();
+
+        /// <summary>
+        /// Does the validation before we send a the Create/Modify request.
+        /// </summary>
+        internal abstract void Validate();
+
+    }
+}
